requirements.txt
torch
torchvision

git+https://github.com/kohya-ss/sd-scripts

## Thiese versions is based on kohya-ss/sd-scripts
accelerate==0.25.0
transformers==4.36.2
diffusers[torch]==0.25.0
ftfy==6.1.1
opencv-python==4.7.0.68
einops==0.7.0
pytorch-lightning==1.9.0
bitsandbytes==0.43.0
prodigyopt==1.0
lion-pytorch==0.0.6
tensorboard
safetensors==0.4.2
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.20.1
# for Image utils
imagesize==1.4.1

# open clip for SDXL
# open-clip-torch==2.20.0
# For logging
rich==13.7.0


## Requirements for Sampling
k-diffusion

## Requirements for finetune/dataset processing
hakuir
imagesize
lycoris-lora==3.0.0.dev4
pytorch-lightning
wandb
warmup-scheduler

## Requirements for minigpt4
torchaudio
matplotlib==3.7.0
psutil==5.9.4
iopath
pyyaml==6.0
regex==2022.10.31
tokenizers
tqdm==4.64.1
timm==0.6.13
webdataset==0.2.48
omegaconf==2.3.0
decord==0.6.0
peft==0.2.0
sentence-transformers
gradio==3.47.1
scikit-image
visual-genome
wandb